refactor: move inliner to optimizer package
